+++ src/app/todo/components/todo.component.html
@@ -3,7 +3,7 @@
     <div class="container-fluid" style="background-color: #303030">
         <div class="row">
             <div style="background-color: #303030;min-height: 100%; min-width: min-content;" class="col-md-3">
-                <app-task-form (task)="addTask($event)"></app-task-form>
+                <app-task-form></app-task-form>
             </div>
             <div style="background-color: #303030;min-height:100%; min-width: min-content;" class="col-md-9 my-3">
                 <app-task-list></app-task-list>
