+++ src/app/todo/components/todo.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';;
+
+@Component({
+  selector: 'app-todo',
+  templateUrl: './todo.component.html',
+  styleUrls: ['./todo.component.scss']
+})
+export class TodoComponent {
+
+}
